detectFiducialMarkers now automatically selects s32 instead of u16, if maxConnectedComponentSegments is above 2^16-1. The unit test output is identical. Note that this may require pushing one of the lists to offchip, which could cause additional speed losses.

+++ robot/test/embeddedVisionTests.cpp
@@ -3200,8 +3200,9 @@ GTEST_TEST(CoreTech_Vision, DetectFiducialMarkers)
   const f32 decode_minContrastRatio = 1.25;
 
   const s32 maxMarkers = 100;
-  //const u16 maxConnectedComponentSegments = 5000; // 25000/4 = 6250
-  const u16 maxConnectedComponentSegments = 39000; // 322*240/2 = 38640
+  //const s32 maxConnectedComponentSegments = 5000; // 25000/4 = 6250
+  const s32 maxConnectedComponentSegments = 39000; // 322*240/2 = 38640
+  //const s32 maxConnectedComponentSegments = 70000;
 
   const s32 quadRefinementIterations = 5;
   const s32 numRefinementSamples = 100;
